requirements.txt
# ETH余额监控工具依赖包
# 核心依赖
web3>=6.0.0
PyYAML>=6.0
requests>=2.28.0

# 可选依赖（用于更好的用户体验）
colorama>=0.4.4  # 彩色终端输出
rich>=13.0.0     # 更好的控制台显示
prettytable>=3.0.0  # 表格显示

# 开发依赖（可选）
pytest>=7.0.0
flake8>=5.0.0
black>=22.0.0
